requirements.txt
defusedxml==0.7.1
fonttools==4.54.1
fpdf2==2.8.1
pillow==11.0.0
pypdf==5.0.1
typing_extensions==4.12.2
